+++ app/src/main/java/com/nekkiichi/edusign/utils/extension/LongExt.kt
@@ -1,4 +1,4 @@
-package com.nekkiichi.edusign.utils
+package com.nekkiichi.edusign.utils.extension
 
 import java.util.Locale
