+++ app/src/main/java/com/nekkiichi/edusign/utils/extension/NavControllerExt.kt
@@ -1,4 +1,4 @@
-package com.nekkiichi.edusign.utils
+package com.nekkiichi.edusign.utils.extension
 
 import androidx.navigation.NavController
 import androidx.navigation.NavOptionsBuilder
